+++ R/TaylorDiagram.R
@@ -255,6 +255,7 @@ TaylorDiagram <- function(
   extra.args$title <- quickText(extra.args$title, auto.text)
   extra.args$subtitle <- quickText(extra.args$subtitle, auto.text)
   extra.args$caption <- quickText(extra.args$caption, auto.text)
+  extra.args$tag <- quickText(extra.args$tag, auto.text)
   extra.args$shape <- extra.args$shape %||% 20
   extra.args$size <- extra.args$size %||% 2
   extra.args$fontsize <- extra.args$fontsize %||% 11
@@ -580,6 +581,7 @@ TaylorDiagram <- function(
       title = extra.args$title,
       subtitle = extra.args$subtitle,
       caption = extra.args$caption,
+      tag = extra.args$tag,
       color = quickText(key.title, auto.text = auto.text),
       shape = quickText(key.title, auto.text = auto.text)
     ) +