+++ R/TheilSen.R
@@ -218,6 +218,7 @@ TheilSen <- function(
   extra.args$title <- quickText(extra.args$title, auto.text)
   extra.args$subtitle <- quickText(extra.args$subtitle, auto.text)
   extra.args$caption <- quickText(extra.args$caption, auto.text)
+  extra.args$tag <- quickText(extra.args$tag, auto.text)
 
   # 2. Time Interval & Data Preparation
   interval <- find_time_interval(mydata$date)
@@ -636,6 +637,7 @@ build_theilsen_plot <- function(
       y = extra.args$ylab,
       title = extra.args$title,
       subtitle = extra.args$subtitle,
-      caption = extra.args$caption
+      caption = extra.args$caption,
+      tag = extra.args$tag
     )
 }