+++ R/percentileRose.R
@@ -227,6 +227,7 @@ percentileRose <- function(
   extra.args$ylab <- quickText(extra.args$ylab, auto.text)
   extra.args$title <- quickText(extra.args$title, auto.text)
   extra.args$subtitle <- quickText(extra.args$subtitle, auto.text)
+  extra.args$tag <- quickText(extra.args$tag, auto.text)
 
   # separate handling for being overwritten
   if ("caption" %in% names(extra.args)) {
@@ -522,7 +523,8 @@ percentileRose <- function(
       y = extra.args$ylab,
       title = extra.args$title,
       subtitle = extra.args$subtitle,
-      caption = extra.args$caption %||% sub
+      caption = extra.args$caption %||% sub,
+      tag = extra.args$tag
     ) +
     ggplot2::scale_colour_manual(
       values = c(