+++ R/trajPlot.R
@@ -306,6 +306,7 @@ trajPlot <- function(
       title = quickText(extra.args$title, auto.text),
       subtitle = quickText(extra.args$subtitle, auto.text),
       caption = quickText(extra.args$caption, auto.text),
+      tag = quickText(extra.args$tag, auto.text),
       colour = quickText(key.title, auto.text)
     )